add dora calculation on uradora (normal dora doesn't need this)
